requirements.txt
absl-py>=0.7.1
cachetools>=3.1.1
certifi>=2019.6.16
chardet>=3.0.4
google-api-core>=1.14.0
google-api-python-client>=1.7.10
google-auth>=1.6.3
google-auth-httplib2>=0.0.3
google-auth-oauthlib>=0.4.0
google-cloud-core>=1.0.2
google-cloud-logging>=1.11.0
googleapis-common-protos>=1.6.0
grpcio>=1.22.0
httplib2>=0.13.0
idna>=2.8
oauthlib>=3.0.2
protobuf>=3.9.0
pyasn1>=0.4.5
pyasn1-modules>=0.2.5
pytz>=2019.1
requests>=2.22.0
requests-oauthlib>=1.2.0
rsa>=4.0
six>=1.12.0
tzlocal>=2.0.0
uritemplate>=3.0.0
urllib3>=1.25.3
